Replacement shutdown commands for Win9x
Just extract this in root / C: of installation after initial setup completes.

Sometimes when using WHPX with Qemu win9x does not shutdown properly causing the system to ask for a system file check. 
With these files the OS will exit or reboot properly. 

This does not override or replace existing reboot and shutdown. These are shortcuts that need to be used or called in order to work.